requirements.txt
tk==0.1.0
tkcalendar==1.6.1
GitPython==3.1.44
